[#942] Moved interpolate_along_los into its own module for cleaner code
